pyproject.toml
[tool.poetry]
name = "telegrambot-1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10.1"
pyTelegramBotAPI = "^3.7.7"
yfinance = "^0.1.59"
openai = "^0.27.7"
langchain = "^0.0.176"
telebot = "^0.0.5"
python-telegram-bot = "^20.3"
setuptools = "^67.8.0"
python-dotenv = "^1.0.0"
emoji = "^2.2.0"
pyyaml = "^6.0"
fastapi = "^0.95.2"
httpx = "^0.24.1"
gunicorn = "^20.1.0"
uvicorn = "^0.22.0"
pydantic = "^1.10.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
